+++ tools/emscripten.py
@@ -1007,6 +1007,7 @@ def create_pointer_conversion_wrappers(metadata):
     '_emval_coro_resume': '_pp',
     'emscripten_main_runtime_thread_id': 'p',
     '_emscripten_set_offscreencanvas_size_on_thread': '_pp__',
+    'fileno': '_p',
   }
 
   for function in settings.SIGNATURE_CONVERSIONS: